+++ pentablake.cu
@@ -493,7 +493,7 @@ extern "C" int scanhash_pentablake(int thr_id, uint32_t *pdata, const uint32_t *
 	uint32_t endiandata[20];
 	int rc = 0;
 	uint32_t throughput = opt_work_size ? opt_work_size : (128 * 2560); // 18.5
-	throughput = min(throughput, (int)(max_nonce - first_nonce));
+	throughput = min(throughput, max_nonce - first_nonce);
 
 	if (opt_benchmark)
 		((uint32_t*)ptarget)[7] = 0x000F;
